Split the interest-order test fixture into its own module
