+++ programs/marginfi/tests/user_actions/mod.rs
@@ -12,6 +12,7 @@ mod emissions;
 mod flash_loan;
 mod indexer_flags;
 mod interest_order;
+mod interest_order_common;
 mod juplend;
 mod kamino;
 mod limit_orders;
